Improved buffer resizing logic
In this update, I resize the memory only when the total bytes written is
greater the buffer size which 1024. Previously, I was resizing based on
the value of n until I realized how much memory that was actually
taking. So for the sake of efficiency and better code, it has been
changed to only reallocate more memory when the buffer size threshold is
encountered.